refactor: extract references code into base component
