+++ packages/api-reference/src/components/ApiClientModal.vue
@@ -7,7 +7,6 @@ import { default as Sidebar } from './Sidebar.vue'
 
 defineProps<{
   parsedSpec: Spec
-  isDarkMode: boolean
   overloadShow?: boolean
   tabMode?: boolean
   activeTab?: string
@@ -46,9 +45,7 @@ export { useApiClientStore } from '@scalar/api-client'
               <div class="t-doc__sidebar">
                 <Sidebar
                   v-show="!isMobile"
-                  :isDarkMode="isDarkMode"
-                  :parsedSpec="parsedSpec"
-                  @toggleDarkMode="$emit('toggleDarkMode')" />
+                  :parsedSpec="parsedSpec" />
               </div>
             </template>
             <template v-else>
@@ -59,9 +56,7 @@ export { useApiClientStore } from '@scalar/api-client'
             <div class="t-doc__sidebar">
               <Sidebar
                 v-show="!isMobile"
-                :isDarkMode="isDarkMode"
-                :parsedSpec="parsedSpec"
-                @toggleDarkMode="$emit('toggleDarkMode')" />
+                :parsedSpec="parsedSpec" />
             </div>
           </template>
           <ApiClient
