+++ packages/api-reference/src/index.ts
@@ -4,6 +4,7 @@ export {
 } from './components/ApiClientModal.vue'
 export { default as ApiReference } from './components/ApiReference.vue'
 export { default as ApiReferenceLayout } from './components/ApiReferenceLayout.vue'
+export { default as ApiReferenceBase } from './components/ApiReferenceBase.vue'
 export { default as Sidebar } from './components/Sidebar.vue'
 export { default as RenderedReference } from './components/Content/Content.vue'
 export { default as DarkModeToggle } from './components/DarkModeToggle.vue'
